% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{failures}
\alias{failures}
\title{Failure log of a cross-validated run}
\usage{
failures(cv)
}
\arguments{
\item{cv}{An object from \code{\link[=cv_predict]{cv_predict()}}.}
}
\value{
A data frame of result column, reason and count.
}
\description{
Failure log of a cross-validated run
}
